Add files via upload
In this new version: 

I have been trying to construct a score that will rank coin pairs by how likely it is that they are duplicates (high match score indicates high likelihood of being a duplicate). The difference in total amount of coins for each coin find in a pair is the most important way to identify whether two coins are duplicates. However, in the past iteration of the program, some coin pairs with differences of ~50-100 were still receiving high match scores. 

I added the total.dif variable to the match score: 

For each pair, I subtract the radius coin find amount from the central coin find amount. Then I calculate where each value falls on a distribution of all these differences. For example, if the greatest difference is 300, then total.dif would be 0, and if the smallest difference is 0, then total.dif would be 1.  All other differences would be some fraction of 1.

I have weighted “total.dif” the most because it is most related to whether the coin find amounts in a pair are equal. I have weighted singleton the least, since when isolating this metric, it caused a positive correlation between match score and coin amounts (I didn’t include the graph here for brevity). The other metrics I’ve weighted equally because they all are not really correlated with the difference in coin amounts.

I also eliminated duplicate pairs that have a difference in coin finds of above 300. I’ve attached a list of the values that got eliminated. This change made a big difference in making the match score more predictive. The large outlier duplicates, some with coin find amount differences of 1000 to 6000, were throwing off metrics that were distribution based like “total.dif.” Duplicate pairs with differences of 50-100 were receiving higher match scores because they were a much better match than those with differences in the thousands.